backend/requirements.txt
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
